Fixed getNeighbours and made updateAngle method kinda better i guess. Author: David

+++ desktop/src/com/mygdx/game/desktop/Sapien.java
@@ -8,7 +8,7 @@ import com.badlogic.gdx.math.Rectangle;
 public abstract class Sapien extends Sprite implements Movable{
     protected TextureAtlas textureAtlas;
 
-    protected int angle;
+    protected float angle;
 
     protected float speed;
     protected int health;
